Time: 4 ms (92.04%), Space: 10 MB (25.35%) - LeetHub

+++ 1464-maximum-product-of-two-elements-in-an-array/1464-maximum-product-of-two-elements-in-an-array.cpp
@@ -4,7 +4,7 @@ public:
         
         sort(nums.begin(),nums.end());
         int n = nums.size();
-        long pro = (nums[n-1] - 1)* (nums[n-2] - 1);
+        int pro = (nums[n-1] - 1)* (nums[n-2] - 1);
         return pro;
     }
 };
